api/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
beanie
python-decouple
passlib
pyjwt
strawberry-graphql[fastapi]